requirements.txt
fastapi
uvicorn
hypercorn
langchain
langchain-community
langchain-google-genai>=0.0.19
langchain-qdrant>=0.1.0
qdrant-client>=1.8.0
python-dotenv
pypdf
bs4
google-cloud-firestore